legacy/requirements.txt
pandas
tqdm
requests
krakenex
ccxt
pyyaml